requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.2
sqlalchemy==2.0.41
pymysql==1.1.1
pydantic-settings==2.9.1
streamlit==1.45.1
requests==2.32.3
openpyxl==3.1.5
xlsxwriter==3.2.0
